requirements.txt
Django==4.2.23
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
django-ninja==1.5.1
pydantic>=2.5.0
python-docx==1.1.0 